requirements.txt
python-docx>=1.1.0
pypdf>=3.0.0
streamlit>=1.28.0
requests>=2.31.0
beautifulsoup4>=4.12.0
google-generativeai>=0.3.0
streamlit-authenticator>=0.2.3
pyyaml>=6.0
st-supabase-connection>=0.1.0
supabase>=2.0.0

